@86erbn3hp #implemented DialogBox, ControlledCheckBox, useModal. Removed comments, separated grids to another file.
