+++ packages/core-library/system/app/internal/blocks/Hub/Settings/SettingsManagement/RefundModal/RefundModalStyles.ts
@@ -10,19 +10,69 @@ export const gridBoxSx: SxProps = {
   borderRight: "2px solid #0F2A714D",
 };
 
-export const modalBoxSx: SxProps = {
-  position: "absolute",
-  top: "50%",
-  left: "50%",
-  transform: "translate(-50%, -50%)",
-  height: "80dvh",
-  maxHeight: "800px",
-  width: "90%",
-  maxWidth: "1400px",
-  borderRadius: "20px",
-  overflow: "hidden",
-  backgroundColor: "#FFF",
-  paddingBottom: "80px",
+export const dialogBoxSx: SxProps = {
+  "& h4": {
+    backgroundColor: "#0F2A71",
+    height: "100px",
+    display: "flex",
+    flexDirection: "row-reverse",
+    alignItems: "center",
+    justifyContent: "space-between",
+    padding: "0 30px",
+    color: "#FFF",
+  },
+
+  "& h4 > span.MuiTypography-root": {
+    fontSize: "32px",
+    fontFamily: "PT Sans",
+  },
+
+  "& [data-testid='close_button_container']": {
+    width: "unset",
+    display: "block",
+  },
+
+  "& .MuiDialog-paper": {
+    height: "80dvh",
+    maxHeight: "800px",
+    width: "90%",
+    maxWidth: "1400px",
+    borderRadius: "20px",
+    overflow: "hidden",
+    paddingBottom: "50px",
+    margin: 0,
+  },
+
+  "& .MuiDialogContent-root": {
+    padding: 0,
+  },
+};
+
+export const submitDialogBoxSx: SxProps = {
+  "& .MuiDialog-paper": {
+    borderRadius: "20px",
+    maxWidth: "900px",
+    height: "unset",
+    minHeight: "450px",
+    padding: "40px",
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "space-between",
+  },
+
+  "& .MuiDialog-paper > .MuiTypography-root": {
+    height: "unset",
+    padding: 0,
+  },
+
+  "& .MuiSvgIcon-root": {
+    color: "#333333",
+  },
+
+  "& .MuiDialogContent-root": {
+    padding: 0,
+    height: "unset",
+  },
 };
 
 export const btnSx: SxProps = {
@@ -40,4 +90,25 @@ export const cardSx = {
   padding: "30px",
   bgcolor: "#0F2A710D",
   borderRadius: "20px",
+};
+
+export const stepperSx: SxProps = {
+  marginX: "auto",
+  gap: "10px",
+  paddingX: "20px",
+  paddingBottom: "40px",
+  justifyContent: "center",
+
+  "& .MuiBox-root:has(.MuiTypography-root)": {
+    width: "unset",
+  },
+
+  "& .MuiStepConnector-root": {
+    width: "100%",
+    maxWidth: "120px",
+  },
+
+  "& .MuiTypography-root": {
+    fontSize: "clamp(12px,2vw,20px)",
+  },
 };
